+++ docs/news.rst
@@ -21,6 +21,7 @@ Psycopg 3.1.9 (unreleased)
   (:ticket:`#543`).
 - Fix loading ROW values with different types in the same query using the
   binary protocol (:ticket:`#545`).
+- Fix dumping recursive composite types (:ticket:`#547`).
 
 
 Current release